backend/requirements.txt
# FastAPI and Server
fastapi
uvicorn[standard]
pydantic-settings

# Database (MongoDB)
beanie
motor

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]
google-auth
requests

# For testing
pytest
httpx
pytest-asyncio

# Generate PDF 
reportlab